+++ Sources/VCore/Services and Managers/SessionManager.swift
@@ -12,16 +12,14 @@ import Foundation
 ///
 /// Object contains `shared` instance, but can also be initialized for separate incrementation.
 ///
-///     let sessionManager: SessionManager = .init()
-///
 ///     Task(operation: {
 ///         do {
-///             let sessionID: Int = sessionManager.newSessionID
+///             let sessionID: Int = await SessionManager.shared.newSessionID
 ///
 ///             let request: NetworkRequest = .init(url: "https://httpbin.org/get")
 ///             let result: [String: Any?] = try await NetworkClient.default.json(from: request)
 ///
-///             guard sessionManager.sessionIsValid(id: sessionID) else { return }
+///             guard await SessionManager.shared.sessionIsValid(id: sessionID) else { return }
 ///
 ///             print(result)
 ///
@@ -30,34 +28,33 @@ import Foundation
 ///         }
 ///     })
 ///
-public final class SessionManager {
+public actor SessionManager {
     // MARK: Properties
-    private let dispatchSemaphore: DispatchSemaphore = .init(value: 1)
+    private var value: Int
     
-    private let atomicInteger: AtomicInteger = .init(initialValue: 0)
-    
-    /// Current session ID. Default to `-1` without any prior incrementation.
-    private(set) public var currentSessionID: Int = -1
-    
-    /// Generates thread-safe, auto-incremented session identifier from `AtomicInteger`.
-    public var newSessionID: Int {
-        dispatchSemaphore.wait()
-        defer { dispatchSemaphore.signal() }
-        
-        currentSessionID = atomicInteger.value
-        return currentSessionID
-    }
-    
-    /// Shared instance of `NetworkSessionManager`.
-    public static let shared: AtomicInteger = .init()
+    /// Shared instance of `SessionManager`.
+    public static let shared: SessionManager = .init()
     
     // MARK: Initializers
-    /// Initializes `SessionManager`.
-    public init() {}
+    /// Initializes `SessionManager` with an initial ID.
+    public init(initialID: Int = 0) {
+        self.value = initialID
+    }
 
-    // MARK: Validation
+    // MARK: ???
+    /// Current session ID.
+    public var currentSessionID: Int {
+        value
+    }
+    
+    /// Increments and returns current session ID.
+    public var newSessionID: Int {
+        value += 1
+        return value
+    }
+    
     /// Validates session identifier against latest generated identifier.
     public func sessionIsValid(id sessionID: Int) -> Bool {
-        sessionID == currentSessionID
+        value == sessionID
     }
 }